requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.2
django-cors-headers==4.3.1
python-decouple==3.8
